testapi/requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=2.0.0,!=2.1.0  # Apache-2.0
setuptools>=16.0,!=24.0.0,!=34.0.0,!=34.0.1,!=34.0.2,!=34.0.3,!=34.1.0,!=34.1.1,!=34.2.0,!=34.3.0,!=34.3.1,!=34.3.2  # PSF/ZPL
tornado>=3.1,<=4.3  # Apache-2.0
epydoc>=0.3.1
six>=1.9.0  # MIT
motor  # Apache-2.0
python-cas
requests[security]
PyJWT